Resolving suggested changes
