+++ package.json
@@ -49,7 +49,7 @@
     "style-loader": "^3.3.1",
     "ts-loader": "^9.4.1",
     "typescript": "^4.8.4",
-    "typescript-plugin-css-modules": "^2.8.0",
+    "typescript-plugin-css-modules": "^3.4.0",
     "url-loader": "^4.1.1",
     "webpack": "^5.74.0",
     "webpack-cli": "^4.10.0",
