+++ src/css-modules/main.module.css
@@ -1,6 +1,7 @@
 .background {
     background-color: #232323;
     margin: 0;
+    padding-bottom: 100px;
 }
 
 .navbar {
