+++ src/redux/store.ts
@@ -2,7 +2,7 @@ import { configureStore } from '@reduxjs/toolkit';
 import reducer from './reducers';
 
 const store = configureStore({
-    reducer: reducer,
+    reducer,
 });
 
 export default store;
