+++ src/utils/camelCase.ts
@@ -1,4 +1,4 @@
-const camelCase = (str) => {
+const camelCase = (str: string) => {
     return str[0].toUpperCase() + str.slice(1, str.length + 1)
 }
