+++ src/utils/timeConverter.ts
@@ -1,8 +1,8 @@
-const timeConverter = (number) => {
+const timeConverter = (number: number) => {
     const hours = Math.floor(number / 60);
     const minutes = number % 60;
 
-    return `${ hours}h ${ minutes }min`;
+    return `${ hours }h ${ minutes }min`; 
 }
 
 export default timeConverter;
